+++ project-documentation-manager.sln
@@ -6,10 +6,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{05882FCC-7708-45B0-9C1A-C02A5054D331}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Business", "src\Business\Business.csproj", "{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Business.Tests", "test\Business.Tests\Business.Tests.csproj", "{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Console", "src\Console\Console.csproj", "{93FC3D1E-2E44-4E1B-8980-707414DCA600}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Console.Tests", "test\Console.Tests\Console.Tests.csproj", "{0E79558A-10ED-4897-B0D1-2042FEBF11F4}"
@@ -22,14 +18,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|Any CPU.Build.0 = Release|Any CPU
 		{93FC3D1E-2E44-4E1B-8980-707414DCA600}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{93FC3D1E-2E44-4E1B-8980-707414DCA600}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93FC3D1E-2E44-4E1B-8980-707414DCA600}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -47,8 +35,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{666F3C4F-CFD5-4DA1-B46E-9253E1A5BFF0} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
-		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890} = {05882FCC-7708-45B0-9C1A-C02A5054D331}
 		{93FC3D1E-2E44-4E1B-8980-707414DCA600} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{0E79558A-10ED-4897-B0D1-2042FEBF11F4} = {05882FCC-7708-45B0-9C1A-C02A5054D331}
 		{19AED20A-034A-49F4-A931-2FDDBE60D3A8} = {05882FCC-7708-45B0-9C1A-C02A5054D331}
