+++ src/Console/Services/MarkdownFileCollectorService.cs
@@ -1,14 +1,13 @@
 using Microsoft.Extensions.Logging;
-using ProjectDocumentationManager.Business.Models;
+using ProjectDocumentationManager.Console.Models;
 
-namespace ProjectDocumentationManager.Business.Services;
+namespace ProjectDocumentationManager.Console.Services;
 
 /// <summary>
 /// Service for collecting markdown source and template files from a directory
 /// </summary>
 public class MarkdownFileCollectorService(ILogger<MarkdownFileCollectorService> logger) : IMarkdownFileCollectorService
 {
-
 
     public async Task<IEnumerable<MarkdownDocument>> CollectAllMarkdownFilesAsync(string directoryPath, CancellationToken cancellationToken = default)
     {
